refactor: 메시지 전송 시 즉시 검증으로 리팩토링
