+++ src/main/java/umc/cockple/demo/domain/chat/service/ChatValidator.java
@@ -1,0 +1,59 @@
+package umc.cockple.demo.domain.chat.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.Request.FileInfo;
+import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.Request.ImageInfo;
+import umc.cockple.demo.domain.chat.exception.ChatErrorCode;
+import umc.cockple.demo.domain.chat.exception.ChatException;
+import umc.cockple.demo.domain.chat.repository.ChatRoomMemberRepository;
+import umc.cockple.demo.domain.chat.repository.ChatRoomRepository;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class ChatValidator {
+
+    private final ChatRoomRepository chatRoomRepository;
+    private final ChatRoomMemberRepository chatRoomMemberRepository;
+
+    public void validateSendRequest(Long chatRoomId, String content, List<FileInfo> files, List<ImageInfo> images, Long senderId) {
+        validateChatRoom(chatRoomId);
+        validateMessage(content, files, images);
+        validateChatRoomMember(chatRoomId, senderId);
+    }
+
+    // ========== 세부 검증 메서드 ==========
+
+    private void validateChatRoom(Long chatRoomId) {
+        if (chatRoomId == null) {
+            throw new ChatException(ChatErrorCode.CHATROOM_ID_NECESSARY);
+        }
+
+        if (!chatRoomRepository.existsById(chatRoomId)) {
+            throw new ChatException(ChatErrorCode.CHAT_ROOM_NOT_FOUND);
+        }
+    }
+
+    private void validateMessage(String content, List<FileInfo> files, List<ImageInfo> images) {
+        boolean hasContent = content != null && !content.trim().isEmpty();
+        boolean hasFiles = files != null && !files.isEmpty();
+        boolean hasImages = images != null && !images.isEmpty();
+
+        if (!hasContent && !hasFiles && !hasImages) {
+            throw new ChatException(ChatErrorCode.EMPTY_MESSAGE_NOT_ALLOWED);
+        }
+
+        if (content != null && content.length() > 1000) {
+            throw new ChatException(ChatErrorCode.MESSAGE_TO_LONG);
+        }
+    }
+
+    private void validateChatRoomMember(Long chatRoomId, Long memberId) {
+        if (!chatRoomMemberRepository.existsByChatRoomIdAndMemberId(chatRoomId, memberId))
+            throw new ChatException(ChatErrorCode.CHAT_ROOM_MEMBER_NOT_FOUND);
+    }
+}
